refactor: cleanup directories & import paths
