+++ web/components/landingpage/effect/index.ts
