+++ web/layouts/landingpage/design-platforms/mobile.tsx
@@ -2,7 +2,7 @@ import Image from "next/image";
 import React, { useState } from "react";
 import styled from "@emotion/styled";
 import { Flex } from "rebass";
-import LiveDesignDemoFrame from "components/motion/live-design-demo";
+import LiveDesignDemoFrame from "components/landingpage/motion/live-design-demo";
 import { media } from "utils/styled/media";
 import { ThemeInterface } from "utils/styled/theme";
