+++ web/sections/landingpage/index.ts
@@ -1,5 +1,4 @@
 import Hero from "./hero";
-import DesignToCode from "./online-app/design-code";
 import OnlineApp from "./online-app";
 import LayoutDetect from "./layout-detect";
 import Products from "./products";
@@ -8,8 +7,7 @@ import Slogan from "./slogan";
 
 export default {
   Hero: Hero,
-  Section2_1_design_to_code: DesignToCode,
-  Section2_2_just_got_online: OnlineApp,
+  Section2_design_to_code: OnlineApp,
   Section3_how_engine_works: LayoutDetect,
   Section4_features_tab: Products,
   Section5_collaboration: Collaborate,
