+++ web/sections/landingpage/online-app/index.tsx
@@ -7,8 +7,8 @@ import { media } from "utils/styled/media";
 import { ThemeInterface } from "utils/styled/theme";
 import ActionItem from "components/action-item";
 import { LandingpageUrls } from "utils/landingpage/constants";
-import OnairButton from "components/effect/onair-button";
-import ApplicationPreview from "layout/application-preview";
+import OnairButton from "components/landingpage/effect/onair-button";
+import ApplicationPreview from "layout/landingpage/application-preview";
 import { DesktopView, MobileView } from "utils/styled/styles";
 import DesignToCode from "./design-code";
 import Image from "next/image";
