+++ web/sections/landingpage/slogan/index.tsx
