+++ web/sections/pricing/index.ts
@@ -1,5 +1,5 @@
 import FreePlan from "./free-plan";
-import PlanList from "./plan-table";
+import PlanList from "./pricing-table";
 import PricingFAQs from "./faqs.section";
 import FeatureList from "./feature-list";
